requirements.txt
# Core framework
Django==5.2.11
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-filter==25.2
drf-spectacular==0.29.0
asgiref==3.11.1
sqlparse==0.5.5
django-cors-headers

# Multi-tenancy & time
django-timezone-field==7.2.1
tzdata==2025.3
tzlocal==5.3.1

# Async / Background jobs
celery==5.6.2
django-celery-beat==2.8.1
redis==7.1.0
amqp==5.3.1
billiard==4.2.4
kombu==5.6.2
vine==5.1.0

# Database
psycopg-binary==3.3.2

# Security / Auth / API utils
PyJWT==2.11.0
uritemplate==4.2.0
inflection==0.5.1

# Environment & config
python-dotenv==1.2.1
PyYAML==6.0.3
packaging==26.0
typing_extensions==4.15.0

# Scheduling
python-crontab==3.3.0
cron_descriptor==2.0.6
python-dateutil==2.9.0.post0

# Validation / Schema
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# CLI / Console utils (safe to keep)
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
prompt_toolkit==3.0.52
wcwidth==0.6.0
six==1.17.0

# Testing (kept intentionally – CI & dev)
pytest==9.0.2
pytest-django==4.12.0
pluggy==1.6.0
iniconfig==2.3.0
Pygments==2.19.2
attrs==25.4.0
celery
redis
